Returns a list of all friends you've ever played
A lot of the basic functionality is working.  Friends will return every
game you've ever played and who every team member was.  You can also
give it another gamertag and it will tell you how many games you've
played with that player.

To do: I need to clean it up a bit and make it a bit more dynamic so you
can pass in what gamertags you want it to compare etc.